requirements.txt
aiocache==0.12.3
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altgraph==0.17.5
asgiref==3.11.0
attrs==25.4.0
bottle==0.13.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
clr_loader==0.2.9
contourpy==1.3.3
cycler==0.12.1
Django==6.0
django-scheduler==0.10.1
fonttools==4.61.0
frozenlist==1.8.0
Hydro-Quebec-API-Wrapper==4.2.5
icalendar==6.3.2
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mplcursors==0.7
multidict==6.7.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.3.251201
pefile==2024.8.26
pillow==12.0.0
pip-review==1.3.0
pkce==1.0.3
propcache==0.4.1
proxy_tools==0.1.0
pycparser==2.23
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2025.2
pywebview==6.1
pywin32-ctypes==0.2.3
requests==2.32.5
schedule==1.2.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
tk==0.1.0
types-pytz==2025.2.0.20251108
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
yarl==1.22.0
